Add static methods to differentiate message from message templates.
Ideally, everything Serilog methods take are message templates, which is great! But more often than not, we do have to end up writing simple text messages, like 

Log.Information("That is complete!");

The problem with this, is that message templates are cached internally, and hence is adds pressure and reduces performance being a different string each time. Use message templates are what is the best practice, but for the simplest of the situations, this end up being counter-intuitive. 

Solution: Pass all text without parameters into the same template "{_}" so that it always uses that specific cached template, and pass the text as a parameter, thereby enriching itself with the ability to pass simple string without the performance hit.