`Serilog.Data` visitor type for dealing with `LogEventPropertyValue` data
Adds a JSON formatter based on this, as a step towards moving sinks away from `JsonFormatter`, which has too many responsibilities and an awkward inheritance-based customisation model.

Adds a rewriter type for modifying (immutable) `LogEventPropertyValue` structures, also based on the visitor. No concrete implementation included, but an example is used in the tests that points the way towards eventual use of this in a payload-size limiting capacity.