Fix AOT problems and introduce feature switch
There's one trim warning that the analyzer didn't catch (bug) that the
AOT compiler does. Unfortunately, the warning isn't easily resolvable --
it implies that you can no longer use deconstructed compiler-generated
types (like anonymous types) through the various logging calls.

There's unfortunately no way to perfectly resolve this. This PR proposes
a behavioral change for trimming -- when trimming, compiler-generated
types will be logged like scalars, instead of being deconstructed. This
is not ideal. Libraries should not have significantly different behavior
when trimming without a warning or some other diagnostic being produced.

Unfortunately, the contract implied by the existing code is much too
complicated to be encoded in the trimmer. The contract is that types do
not need to preserve reflection metadata, except for compiler-generated
types, which need public properties. However, the definition of
compiler-generated types is not well-defined. Encoding a heuristic in
Serilog is one thing -- encoding the heuristic in the .NET runtime is
another.

Given the constraints, this seems like the best option.