fix: allows all printable characters as format values
The existing implementation used .IsPunctuation + IsLetterOrDigit with
an override for '+' and ' '. That left some conspicuous holes and the
spec itself is just "^\}" so basically anything goes in a format string.
This allows for them and this does seem to have a reasonably large
impact on the DefaultConsoleOutputTemplate parsing method. Everything
else was within error bounds and no impact at all on memory allocs.