Fix AOT compatibility (#2175)
* Fix AOT compatibility

It looks like the code has drifted a bit and disrupted the feature switch that
provided AOT compatibility. I've fixed the feature switch, added the AotTestApp
to the solution to ensure that it builds when the solution does, configured the
AotTestApp to publish on build so that it runs the IL analysis engine, and added
a Serilog.props file to the NuGet package to enable the feature switch automatically
whenever trimming is enabled.

This should be much more likely to catch any regressions.

* Fix comment